refactor: AuthManager 패키지 경로 수정 및 불필요한 import 제거
- `AuthManager`가 로컬 데이터를 처리하는 역할을 명확히 하기 위해 패키지 경로를 `data.remote`에서 `data.local`로 변경했습니다.
- `SplashScreen`, `SignInScreen`, `SignUpScreen`, `MyPageScreen`, `MainActivity`에서 변경된 `AuthManager`의 import 경로를 수정했습니다.
- `SignInScreen`에서 사용되지 않는 `LaunchedEffect` import 문을 제거했습니다.
