+++ app/src/main/java/com/sopt/dive/presentation/search/model/SearchType.kt
